fix(backup): if backup path is `content://`, use default data dir
Needed for #https://github.com/status-im/status-desktop/issues/19082

The issue is that on Android, we cannot write directly to `content://`, we need to use the Android Filesystem. So to remedy this, we save to the data dir, which is inside the sandbox, so always available.
We then return that path.

It's the job of the client afterwards to copy that file to the wanted folder using the SAF